Remove duplicated register logic
