+++ packages/atlas/src/components/_auth/SignUpModal/SignUpModal.types.ts
@@ -1,5 +1,3 @@
-import { ImageInputFile } from '@/components/_inputs/MultiFileSelect'
-
 export enum SignUpSteps {
   CreateMember,
   SignUpSeed,
@@ -7,20 +5,4 @@ export enum SignUpSteps {
   SignUpPassword,
   Creating,
   Success,
-}
-
-export type MemberFormData = {
-  handle: string
-  avatar?: ImageInputFile
-  captchaToken?: string
-  mnemonic: string
-  confirmedCopy: boolean
-}
-
-export type AccountFormData = {
-  email: string
-  password: string
-  mnemonic: string
-  confirmedTerms: boolean
-  memberId: string
 }
